pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Paul John <rainson.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
requests = "^2.32.3"
pydantic-settings = "^2.3.4"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
gunicorn = "^22.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
